build: Remove autogen.sh to prevent jenkins-obs confusion
The jenkins-obs script looks for the existence of autogen.sh to run
`make distcheck` to generate a release tarball for the module. However,
the autotooling in PyNDN2 doesn’t include the Python modules (which are
what we actually want) in its tarball.

Instead, we want a tarball generated using setup.py, which is what
jenkins-obs checks for next; delete autogen.sh so it skips straight to
that.

This patch cannot be upstreamed.

Signed-off-by: Philip Withnall <withnall@endlessm.com>